+++ SandboxiePlus/MiscHelpers/Common/Finder.cpp
@@ -114,6 +114,13 @@ CFinder::CFinder(QObject* pFilterTarget, QWidget *parent, int iOptions)
 	pSpacer->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Expanding);
 	m_pSearchLayout->addWidget(pSpacer);
 
+	m_pProgressBar = new QProgressBar(this);
+	m_pProgressBar->setMaximumWidth(150);
+	m_pProgressBar->setMaximumHeight(16);
+	m_pProgressBar->setTextVisible(true);
+	m_pProgressBar->setVisible(false);
+	m_pSearchLayout->addWidget(m_pProgressBar);
+
 	QToolButton* pClose = new QToolButton(this);
     pClose->setIcon(QIcon(":/close.png"));
     pClose->setAutoRaise(true);
@@ -343,7 +350,7 @@ next_parent:
 	return QModelIndex();
 }
 
-void CFinder::OnSelectNext() 
+void CFinder::OnSelectNext()
 {
 	if (!m_pModel)
 		return;
@@ -354,4 +361,21 @@ void CFinder::OnSelectNext()
 		m_pTree->setCurrentIndex(idx);
 	else
 		QApplication::beep();
+}
+
+void CFinder::SetProgress(int value, int maximum)
+{
+	m_pProgressBar->setMaximum(maximum);
+	m_pProgressBar->setValue(value);
+}
+
+void CFinder::ShowProgress()
+{
+	m_pProgressBar->setValue(0);
+	m_pProgressBar->setVisible(true);
+}
+
+void CFinder::HideProgress()
+{
+	m_pProgressBar->setVisible(false);
 }